Ex A - Step 6
With the data in arrays in ComputeTotal function,
a) If there are no items with qty>0, then we show the message “No food item selected”, and grey out/disable the submit button.
b) If there is at least 1 item with qty>0, we will process those items, and display the details of the items’ name, price and qty. Then we will compute the sub-total, GST and TOTAL amount and also display them. Lastly we will enable/ungrey the submit button.